Ex 4.21

+++ part4/blogilista/controllers/blogs.js
@@ -51,8 +51,22 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({error: 'token invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        await Blog.deleteOne({id: blog.id})
+        response.status(204).end()
+    }
+    return response.status(401).json({error: 'token invalid'})
+
 })
 
 module.exports = blogsRouter
